Refactor service providers and move settings to theme.json
