--- web/app/themes/wp-gds-theme/app/Providers/EditorServiceProvider.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace App\Providers;
-
-use Roots\Acorn\ServiceProvider;
-
-use function Roots\config;
-
-class EditorServiceProvider extends ServiceProvider
-{
-    /**
-     * Register any application services.
-     *
-     * @return void
-     */
-    public function register()
-    {
-        add_action('after_setup_theme', [$this, 'themeSetup']);
-    }
-
-    public function themeSetup(): void
-    {
-        add_theme_support('align-wide');
-        add_theme_support('editor-styles');
-        add_theme_support('wp-block-styles');
-        // add_theme_support('disable-custom-colors');
-        add_theme_support('disable-custom-font-sizes');
-        add_theme_support('disable-custom-gradients');
-        add_theme_support('disable-custom-line-height');
-
-        // add_theme_support('experimental-custom-spacing');
-        // add_theme_support('experimental-link-color');
-
-        add_theme_support('editor-color-palette', config('theme.editor-color-palette'));
-        add_theme_support('editor-font-sizes', config('theme.editor-font-sizes'));
-        add_theme_support('editor-gradient-presets', config('theme.editor-gradient-presets'));
-
-        // Remove core block patterns.
-        // remove_theme_support('core-block-patterns');
-    }
-}
